category index template and controller restructured
