+++ src/Config/Category.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Config;
+
+enum Category: string
+{
+    case Lesson = 'Cours';
+    case Tutoriel = 'Tutoriel';
+    case LiveCoding = 'LiveCoding';
+}
