+++ templates/category/index.html.twig
@@ -10,13 +10,13 @@
 	<div id="index-category">
 		<h1>//CATÉGORIES_</h1>
 		<div id="index-category-cards">
-			{% for category in categories %}
+			{% for key, category in videos %}
 				<div class="category-carousel">
-					<h2>{{category.label}}</h2>
-					<section class="splide" id="{{category.label}}-slider" aria-label="Splide Basic HTML Example">
+					<h2>{{key}}</h2>
+					<section class="splide" id="{{key}}-slider" aria-label="Splide Basic HTML Example">
 						<div class="splide__track">
 							<ul class="splide__list">
-								{% for video in videos[category.label] %}
+								{% for video in category %}
 									<li class="splide__slide">
 										<div class="{{slug}}-card category-card index-card">
 											<div class="recto-index-card">
@@ -38,8 +38,8 @@
 							</ul>
 						</div>
 					</section>
-					<a class="all-videos" href="{{ path("video_show_by_category", { 'categoryLabel': category.label, 'languageSlug': slug })}}">Toutes les vidéos</a>
-					<a class="all-videos-mobile btn-glitch-fill" href="{{ path("video_show_by_category", { 'categoryLabel': category.label, 'languageSlug': slug })}}">
+					<a class="all-videos" href="{{ path("video_show_by_category", { 'categoryLabel': key, 'languageSlug': slug })}}">Toutes les vidéos</a>
+					<a class="all-videos-mobile btn-glitch-fill" href="{{ path("video_show_by_category", { 'categoryLabel': key, 'languageSlug': slug })}}">
 						<span class="text">//Vidéos</span>
 						<span class="text-decoration">_</span>
 						<span class="decoration">&rArr;</span>
