+++ templates/language/index.html.twig
@@ -18,7 +18,7 @@
 					<div class="verso-index-card">
 						<h2>//{{language.label}}_</h2>
 						<p>{{language.description}}</p>
-						<a class="btn-glitch-fill" href="{{ path("category_index", { slug: language.slug })}}">
+						<a class="btn-glitch-fill" href="{{ path("language_category_index", { slug: language.slug })}}">
 							<span class="text">//Catégories</span>
 							<span class="text-decoration">_</span>
 							<span class="decoration">&rArr;</span>
